requirements.txt
torch==2.2.1
torchvision==0.17.1
datasets==2.19.1
transformers==4.40.2
Pillow==9.4.0
accelerate==0.30.1
scikit-learn
seaborn
numpy
matplotlib